+++ src/app/ai-chat/roadmap/page.tsx
@@ -1,3 +1,14 @@
-export default function AIChatRoadMapPage() {
-  return <div>AI채팅 로드맵 페이지</div>;
+'use client';
+
+import { useEffect } from 'react';
+import { useRouter } from 'next/navigation';
+
+export default function AiChatRoadmap() {
+  const router = useRouter();
+
+  useEffect(() => {
+    router.replace('/ai-chat?chapter=roadmap');
+  }, [router]);
+
+  return null;
 }
